proxmox/requirements.txt
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-semantic-conventions==0.46b0
psutil>=5.9.0
pysmart>=1.0.0
requests>=2.28.1
proxmoxer>=1.3.1
pyjwt>=2.6.0
python-dateutil>=2.8.2
pygtail==0.14.0